+++ test/test_reference_signatures.ml
@@ -210,7 +210,8 @@ let documented_operations () =
          in
          (match unquote family with
           | Some fam
-            when List.mem fam ["Shell"; "FS"; "Env"; "IO"; "Proc"; "Clock"] ->
+            when List.mem fam
+                   ["Shell"; "FS"; "Env"; "IO"; "Proc"; "Clock"; "Random"] ->
             List.iter (fun cell ->
               List.iter (fun verb ->
                 match unquote verb with